Cache up to three cairo_image_surfaces
If we show the inline display in the generic UI, in the mixer and in the editor
mixer strip, i.e. in three different sizes, we'd need to constantly reallocate
the cairo_image_surface. Therefore we cache up to three cairo_image_surfaces in
order to reuse them, when a specific display size is requested.

The number of three can be easily extended, if needed.