Fix error when $PATH contains whitespaces
Previously, if $PATH contains whitespaces the startup script would split
the new exported $PATH on the first whitespace. This was observed on an
Ubuntu system where $PATH contained a directory with whitespaces.

The change adds weak quotation around the new path to prevent expansion
of whitespaces. $PATH though, is still expanded.