Consolidated a-comp/a-exp run_mono() and run_stereo() to one run()
Up to now we had in a-comp and a-exp one run_mono() and one run_stero()
function which where almost identical except that run_stereo() treated two
in/outs and run_mono() only one. Now we store the number of channels
acomp->n_channels according to the URI and in run() we loop over an array of
pointers to the in/out buffers.