Reset write-source only when necessary
When I/O port-counts do not change, plugin re-order happens
in sync in the process-thread. ::configure_io() is only called
to ensure that the current configuration is valid.

In case that the ChanCount does not change, the method must
be realtime-safe and not block.

DiskWriter::reset_write_sources() is not realtime-safe and
implicitly causes a session-save:

Write-sources are destroyed and re-created. This includes
a call to write_source->drop_references(), which triggers
ARDOUR::Session::remove_source(), which saves the session.

Furthermore adding/removing plugins will likewise call
::configure_io().

Previously any processor change on a track lead to
saving the session!