Foldback gui: desensitize solo button when in SIP mode
Because foldback by design goes to physical outputs
rather than master or monitor, SIP solo does not work
or make sense. PFL/AFL allow a foldback to be rough
adjusted using the control room monitoring.

A soloed Foldback bus will be released when the foldback
strip is switched to a different foldback bus