Fix thinko in cd5f70e69a (delayline buffer resize)
The delay-time itself can change arbitrarily, but the buffer-size
never shrinks.

If the buffersize grows it means that the new delay is longer than
the current one (or at least as large as any pending, not yet
active delay).

This is important for the mechanism that adjusts the read-pointer
to the new buffer-size.