Leave processor box sensitive when editing aux-sends
This is required to toggle the strip to revert from aux-view
to the main route using the aux-send processor itself.

Discuss:
Perhaps this should be reverted, and ProcessorBox::edit_aux_send()
be removed: It is currently possible to create various
inconsistent mixer-states.

* use "Aux" on a Bus, then double-click some Internal-Send to
  revert strips. But Aux remains active
* Removing sends while "Aux" is active does not disable Aux.

If the processor box was insensitive those issues won't be present.
However it is convenient to use double-click on the editor-mixer to
access a send's panner ...