Fix loading send gain automation
This cleans up various issues that have been accumulated over the past
and fixes bugs introduced in d4e023e1 and e31f5d99.

Previously GainControl as saved as part of the Amp, however the
automation was saved via Send (is-a Automatable).

In d4e023e1cb, the GainControl was changed a "BusSendLevel"
parameter, but AutomationList was not updated. This prevented
loading existing automation (control parameter was not found).