Exclude zita-convolver from being available as external lib
Ardour statically links libardour against zita-convolver,
this is required because it's used in plugins and available
by Lua scripts.

libzita-convolver has been modified to be in a custom namespace
and extern global symbols have been removed.

On most GNU/Linux distros system-wide default zita-convolver
is not compatible nor suitable to be used as-is.